// types.ts
export interface Lesson {
    auditorium: string;
    building: string;
    beginLesson: string;
    endLesson: string;
    dayOfWeekString: string;
    discipline: string;
    kindOfWork: string;
    lecturer: string;
    lessonNumberStart: number;
    date: string;
}

export interface SearchGroup {
    id: string;
    label: string;
    description: string;
}